Aggregation Layer

The aggregation layer of our basic ICN can be constructed using aggregation_layer().

CompositionalNetworks.aggregation_layer — Function

aggregation_layer()

Generate the layer of aggregations of the ICN. The operations are mutually exclusive, that is only one will be selected.

List of Aggregations

Follows a list of the current operations available in any aggregation layer. Those operations are mutually exclusive.

CompositionalNetworks._ag_sum — Function

_ag_sum(x)

Aggregate through + a vector into a single scalar.

CompositionalNetworks._ag_count_positive — Function

_ag_count_positive(x)

Count the number of strictly positive elements of x.