Comparison Layer

The comparison layer of our basic ICN can be constructed using comparison_layer(param=false). All operations are mutually exclusive.

CompositionalNetworks.comparison_layer — Function

comparison_layer(param = false)

Generate the layer of transformations functions of the ICN. Iff param value is set, also includes all the parametric comparison with that value. The operations are mutually exclusive, that is only one will be selected.

Non-parametric comparisons

Follows a list of the current non-parametric operations available in any comparison layer.

Missing docstring.

Missing docstring for CompositionalNetworks._co_identity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._co_euclidian. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._co_abs_diff_val_vars. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._co_val_minus_vars. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._co_vars_minus_val. Check Documenter's build log for details.

Parametric comparisons

And finally a list of the parametric ones.

Missing docstring.

Missing docstring for CompositionalNetworks._co_abs_diff_val_param. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._co_val_minus_param. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._co_param_minus_val. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._co_euclidian_param. Check Documenter's build log for details.