ICNs

Currently only construct a generic ICN with all available operations.

CompositionalNetworks.ICN — Type

ICN(; nvars, dom_size, param, transformation, arithmetic, aggregation, comparison)

Construct an Interpretable Compositional Network, with the following arguments:

nvars: number of variable in the constraint

dom_size: maximum domain size of any variable in the constraint

param: optional parameter (default to nothing)

transformation: a transformation layer (optional)

arithmetic: a arithmetic layer (optional)

aggregation: a aggregation layer (optional)

comparison: a comparison layer (optional)