Transformation Layer

The transformation layer of our basic ICN can be constructed using transformation_layer(param=false).

CompositionalNetworks.transformation_layer — Function

transformation_layer(param = false)

Generate the layer of transformations functions of the ICN. Iff param value is true, also includes all the parametric transformations.

Non-parametric transformations

Follows a list of the current non-parametric operations available in any transformation layer.

CompositionalNetworks._tr_identity — Function

_tr_identity(x)
_tr_identity(i, x)

Identity function. Already defined in Julia as identity, specialized for vectors.

CompositionalNetworks._tr_count_eq — Function

_tr_count_eq(i, x)
_tr_count_eq(x)

Count the number of elements equal to x[i]. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_count_eq_left — Function

_tr_count_eq_left(i, x)
_tr_count_eq_left(x)

Count the number of elements to the left of and equal to x[i]. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_count_eq_right — Function

_tr_count_eq_right(i, x)
_tr_count_eq_right(x)

Count the number of elements to the right of and equal to x[i]. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_count_greater — Function

_tr_count_greater(i, x)
_tr_count_greater(x)

Count the number of elements greater than x[i]. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_count_lesser — Function

_tr_count_lesser(i, x)
_tr_count_lesser(x)

Count the number of elements lesser than x[i]. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_count_g_left — Function

_tr_count_g_left(i, x)
_tr_count_g_left(x)

Count the number of elements to the left of and greater than x[i]. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_count_l_left — Function

_tr_count_l_left(i, x)
_tr_count_l_left(x)

Count the number of elements to the left of and lesser than x[i]. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_count_g_right — Function

_tr_count_g_right(i, x)
_tr_count_g_right(x)

Count the number of elements to the right of and greater than x[i]. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_count_l_right — Function

_tr_count_l_right(i, x)
_tr_count_l_right(x)

Count the number of elements to the right of and lesser than x[i]. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_contiguous_vals_minus — Function

_tr_contiguous_vals_minus(i, x)
_tr_contiguous_vals_minus(x)

Return the difference x[i] - x[i + 1] if positive, 0.0 otherwise. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_contiguous_vals_minus_rev — Function

_tr_contiguous_vals_minus_rev(i, x)
_tr_contiguous_vals_minus_rev(x)

Return the difference x[i + 1] - x[i] if positive, 0.0 otherwise. Extended method to vector with sig (x) are generated.

Note that all functions are extended to a vectorized version with the lazy function.

CompositionalNetworks.lazy — Function

lazy(funcs::Function...)

Generate methods extended to a vector instead of one of its components. A function f should have the following signature: f(i::Int, x::V; param = nothing).

Parametric transformations

And finally a list of the parametric ones.

CompositionalNetworks._tr_count_eq_param — Function

_tr_count_eq_param(i, x; param)
_tr_count_eq_param(x; param)

Count the number of elements equal to x[i] + param. Extended method to vector with sig (x, param) are generated.

CompositionalNetworks._tr_count_l_param — Function

_tr_count_l_param(i, x; param)
_tr_count_l_param(x; param)

Count the number of elements lesser than x[i] + param. Extended method to vector with sig (x, param) are generated.

CompositionalNetworks._tr_count_g_param — Function

_tr_count_g_param(i, x; param)
_tr_count_g_param(x; param)

Count the number of elements greater than x[i] + param. Extended method to vector with sig (x, param) are generated.

CompositionalNetworks._tr_count_bounding_param — Function

_tr_count_bounding_param(i, x; param)
_tr_count_bounding_param(x; param)

Count the number of elements bounded (not strictly) by x[i] and x[i] + param. An extended method to vector with sig (x, param) is generated.

CompositionalNetworks._tr_val_minus_param — Function

_tr_val_minus_param(i, x; param)
_tr_val_minus_param(x; param)

Return the difference x[i] - param if positive, 0.0 otherwise. Extended method to vector with sig (x, param) are generated.

CompositionalNetworks._tr_param_minus_val — Function

_tr_param_minus_val(i, x; param)
_tr_param_minus_val(x; param)

Return the difference param - x[i] if positive, 0.0 otherwise. Extended method to vector with sig (x, param) are generated.

Note that all functions are extended to a vectorized version with the lazy_param function.

CompositionalNetworks.lazy_param — Function

lazy_param(funcs::Function...)

Generate methods extended to a vector instead of one of its components. A function f should have the following signature: f(i::Int, x::V; param).