Compositional Networks

CompositionalNetworks.ICN

CompositionalNetworks._ag_count_positive

CompositionalNetworks._ag_sum

CompositionalNetworks._ar_sum

CompositionalNetworks._as_bitvector

CompositionalNetworks._as_int

CompositionalNetworks._co_abs_diff_val_param

CompositionalNetworks._co_abs_diff_val_vars

CompositionalNetworks._co_euclidian_param

CompositionalNetworks._co_identity

CompositionalNetworks._co_val_minus_param

CompositionalNetworks._co_val_minus_vars

CompositionalNetworks._compose

CompositionalNetworks._exclu

CompositionalNetworks._functions

CompositionalNetworks._generate_inclusive_operations

CompositionalNetworks._generate_population

CompositionalNetworks._generate_weights

CompositionalNetworks._is_viable

CompositionalNetworks._layers

CompositionalNetworks._length

CompositionalNetworks._length

CompositionalNetworks._loss

CompositionalNetworks._map_tr

CompositionalNetworks._nbits

CompositionalNetworks._nbits_exclu

CompositionalNetworks._optimize!

CompositionalNetworks._reduce_symbols

CompositionalNetworks._selected_size

CompositionalNetworks._show_layer

CompositionalNetworks._single_point

CompositionalNetworks._symbol

CompositionalNetworks._tr_contiguous_vals_minus

CompositionalNetworks._tr_count_bounding_param

CompositionalNetworks._tr_count_eq

CompositionalNetworks._tr_count_eq_param

CompositionalNetworks._tr_count_greater

CompositionalNetworks._tr_identity

CompositionalNetworks._tr_val_minus_param

CompositionalNetworks._weigths

CompositionalNetworks._weigths!

CompositionalNetworks.aggregation_layer

CompositionalNetworks.arithmetic_layer

CompositionalNetworks.comparison_layer

CompositionalNetworks.compose

CompositionalNetworks.csv2space

CompositionalNetworks.hamming

CompositionalNetworks.lazy

CompositionalNetworks.regularization

CompositionalNetworks.show_composition

CompositionalNetworks.show_layers

CompositionalNetworks.transformation_layer

CompositionalNetworks.ICN — Type

ICN(; nvars, dom_size, param, transformation, arithmetic, aggregation, comparison)

Construct an Interpretable Compositional Network, with the following arguments:

nvars: number of variable in the constraint

dom_size: maximum domain size of any variable in the constraint

param: optional parameter (default to nothing)

transformation: a transformation layer (optional)

arithmetic: a arithmetic layer (optional)

aggregation: a aggregation layer (optional)

comparison: a comparison layer (optional)

CompositionalNetworks._ag_count_positive — Method

_ag_count_positive(x)

Count the number of strictly positive elements of x.

CompositionalNetworks._ag_sum — Method

_ag_sum(x)

Aggregate through + a vector into a single scalar.

CompositionalNetworks._ar_sum — Method

_ar_sum(x)
_ar_prod(x)

Reduce k = length(x) vectors through sum/product to a single vector.

CompositionalNetworks._as_bitvector — Function

_as_bitvector(n::Int, max_n::Int = n)

Convert an Int to a BitVector of minimal size (relatively to max_n).

CompositionalNetworks._as_int — Method

_as_int(v::AbstractVector)

Convert a BitVector into an Int.

CompositionalNetworks._co_abs_diff_val_param — Method

_co_abs_diff_val_param(x, param)

Return the absolute difference between x and param.

CompositionalNetworks._co_abs_diff_val_vars — Method

_co_abs_diff_val_vars(x, nvars)

Return the absolute difference between x and the number of variables nvars.

CompositionalNetworks._co_euclidian_param — Method

_co_euclidian_param(x, param, ds)
_co_euclidian(x, ds)

Compute an euclidian norm with domain size ds, possibly weigthed by param, on a scalar.

CompositionalNetworks._co_identity — Method

_co_identity(x::Number)

Identity function. Already defined in Julia as identity, specialized for scalars in the comparison layer.

CompositionalNetworks._co_val_minus_param — Method

_co_val_minus_param(x, param)
_co_param_minus_val(x, param)

Return the difference x - param (resp. param - x) if positive, 0.0 otherwise.

CompositionalNetworks._co_val_minus_vars — Method

_co_val_minus_vars(x, nvars)
_co_vars_minus_val(x, nvars)

Return the difference x - nvars (resp. nvars - x) if positive, 0.0 otherwise, where nvars denotes the numbers of variables.

CompositionalNetworks._compose — Method

_compose(icn)

Internal function called by compose and show_composition.

CompositionalNetworks._exclu — Method

_exclu(layer)

Return true if the layer has mutually exclusive operations.

CompositionalNetworks._functions — Method

_functions(layer)

Access the operations of a layer. The container is ordered.

CompositionalNetworks._generate_inclusive_operations — Method

_generate_inclusive_operations(predicate, bits)
_generate_exclusive_operation(max_op_number)

Generates the operations (weigths) of a layer with inclusive/exclusive operations.

CompositionalNetworks._generate_population — Method

_generate_population(icn, pop_size

Generate a pôpulation of weigths (individuals) for the genetic algorithm weigthing icn.

CompositionalNetworks._generate_weights — Method

_generate_weights(layers)
_generate_weights(icn)

Generate the weigths of a collection of layers or of an ICN.

CompositionalNetworks._is_viable — Method

_is_viable(layer, w)
_is_viable(icn)
_is_viable(icn, w)

Assert if a pair of layer/icn and weigths compose a viable pattern. If no weigths are given with an icn, it will check the current internal value.

CompositionalNetworks._layers — Method

_layers(icn)

Return the ordered layers of an ICN.

CompositionalNetworks._length — Method

_length(layer)

Return the number of operations in a layer.

CompositionalNetworks._length — Method

_length(icn)

Return the total number of operations of an ICN.

CompositionalNetworks._loss — Method

_loss(X, X_sols, icn, weigths, metric)

Compute the loss of icn.

CompositionalNetworks._map_tr — Method

_map_tr(f, x)
_map_tr(f, x, param)

Return an anonymous function that applies f to all elements of x, with an optional parameter param.

CompositionalNetworks._nbits — Method

_nbits(icn)

Return the expected number of bits of a viable weigth of an ICN.

CompositionalNetworks._nbits_exclu — Method

_nbits_exclu(layer)

Convert the length of an exclusive layer into a number of bits.

CompositionalNetworks._optimize! — Method

_optimize!(icn, X, X_sols; ga = GA(), metric = hamming, pop_size = 200)

Optimize and set the weigths of an ICN with a given set of configuration X and solutions X_sols.

CompositionalNetworks._reduce_symbols — Function

_reduce_symbols(symbols, sep)

Produce a formatted string that separates the symbols by sep. Used internally for show_composition.

CompositionalNetworks._selected_size — Method

_selected_size(layer, layer_weights)

Return the number of operations selected by layer_weights in layer.

CompositionalNetworks._show_layer — Method

_show_layer(layer)

Return a string that contains the elements in a layer.

CompositionalNetworks._single_point — Method

_single_point(v1::T, v2::T, icn) where {T <: AbstractVector}
_flip(recombinant::T, icn) where {T <: BitVector}

Meta functions that add individuals a viability check in Evolutionary.singlepoint and Evolutionary.flip.

CompositionalNetworks._symbol — Method

_symbol(layer, i)

Return the i-th symbols of the operations in a given layer.

CompositionalNetworks._tr_contiguous_vals_minus — Method

_tr_contiguous_vals_minus(i, x)
_tr_contiguous_vals_minus_rev(i, x)

Return the difference x[i] - x[i + 1] (resp. x[i + 1] - x[i]) if positive, 0.0 otherwise. Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_count_bounding_param — Method

_tr_count_bounding_param(i, x, param)

Count the number of elements bounded (not strictly) by x[i] and x[i] + param. An extended method to vector with sig (x, param) is generated.

CompositionalNetworks._tr_count_eq — Method

_tr_count_eq(i, x)
_tr_count_eq_right(i, x)
_tr_count_eq_left(i, x)

Count the number of elements equal to x[i] (optionally to the right/left of x[i]). Extended method to vector x are generated.

CompositionalNetworks._tr_count_eq_param — Method

_tr_count_eq_param(i, x, param)
_tr_count_l_param(i, x, param)
_tr_count_g_param(i, x, param)

Count the number of elements equal to (resp. lesser/greater than) x[i] + param. Extended method to vector with sig (x, param) are generated.

CompositionalNetworks._tr_count_greater — Method

_tr_count_greater(i, x)
_tr_count_lesser(i, x)
_tr_count_g_left(i, x)
_tr_count_l_left(i, x)
_tr_count_g_right(i, x)
_tr_count_l_right(i, x)

Count the number of elements greater/lesser than x[i] (optionally to the left/right of x[i]). Extended method to vector with sig (x) are generated.

CompositionalNetworks._tr_identity — Method

_tr_identity(x)
_tr_identity(x)

Identity function. Already defined in Julia as identity, specialized for vectors and scalars.

CompositionalNetworks._tr_val_minus_param — Method

_tr_val_minus_param(i, x, param)
_tr_param_minus_val(i, x, param)

Return the difference x[i] - param (resp. param - x[i]) if positive, 0.0 otherwise. Extended method to vector with sig (x, param) are generated.

CompositionalNetworks._weigths! — Method

_weights!(icn, weights)

Set the weights of an ICN with a BitVector.

CompositionalNetworks._weigths — Method

_weigths(icn)

Access the current set of weigths of an ICN.

CompositionalNetworks.aggregation_layer — Method

aggregation_layer()

Generate the layer of aggregation functions of the ICN.

CompositionalNetworks.arithmetic_layer — Method

arithmetic_layer()

Generate the layer of arithmetic functions of the ICN.

CompositionalNetworks.comparison_layer — Function

comparison_layer(nvars, dom_size, param = nothing)

Generate the layer of transformations functions of the ICN. Iff param value is set, also includes all the parametric transformation with that value.

CompositionalNetworks.compose — Method

compose(icn)
compose(icn, weights)

Return a function composed by some of the operations of a given ICN. Can be applied to any vector of variables. If weights are given, will assign to icn.

CompositionalNetworks.csv2space — Method

csv2space(file; filter=:none)

Convert a csv file into a collection of configurations. If the filter is set to :concept, only solutions will be extracted.

CompositionalNetworks.hamming — Method

hamming(x, X)

Compute the hamming distance of x over a collection of solutions X, i.e. the minimal number of variables to switch in xto reach a solution.

CompositionalNetworks.lazy — Method

lazy(funcs::Function...)
lazy_param(funcs::Function...)

Generate methods extended to a vector instead of one of its components. For lazy (resp. lazy_param) a function f should have the following signature: f(i::Int, x::V) (resp. f(i::Int, x::V, param::T)).

CompositionalNetworks.regularization — Method

regularization(icn)

Return the regularization value of an ICN weights, which is proportional to the normalized number of operations selected in the icn layers.

CompositionalNetworks.show_composition — Method

show_composition(icn)

Return the composition (weights) of an ICN.

CompositionalNetworks.show_layers — Method

show_layers(icn)

Return a formated string with each layers in the icn.

CompositionalNetworks.transformation_layer — Function

transformation_layer(param = nothing)

Generate the layer of transformations functions of the ICN. Iff param value is set, also includes all the parametric transformation with that value.