Compositional Networks

CompositionalNetworks.ICN

CompositionalNetworks._abs_diff_val_param

CompositionalNetworks._abs_diff_val_vars

CompositionalNetworks._as_bitvector

CompositionalNetworks._as_int

CompositionalNetworks._contiguous_vals_minus

CompositionalNetworks._count_bounding_param

CompositionalNetworks._count_eq

CompositionalNetworks._count_eq_param

CompositionalNetworks._count_greater

CompositionalNetworks._count_positive

CompositionalNetworks._euclidian_param

CompositionalNetworks._exclu

CompositionalNetworks._functions

CompositionalNetworks._identity

CompositionalNetworks._layers

CompositionalNetworks._length

CompositionalNetworks._length

CompositionalNetworks._map_tr

CompositionalNetworks._nbits_exclu

CompositionalNetworks._sum

CompositionalNetworks._sum

CompositionalNetworks._symbol

CompositionalNetworks._val_minus_param

CompositionalNetworks._val_minus_param

CompositionalNetworks._val_minus_vars

CompositionalNetworks._weigths

CompositionalNetworks.aggregation_layer

CompositionalNetworks.arithmetic_layer

CompositionalNetworks.comparison_layer

CompositionalNetworks.compose

CompositionalNetworks.lazy

CompositionalNetworks.transformation_layer

CompositionalNetworks.ICN — Type

ICN(; nvars, dom_size, param, transformation, arithmetic, aggregation, comparison)

Construct an Interpretable Compositional Network, with the following arguments:

nvars: number of variable in the constraint

dom_size: maximum domain size of any variable in the constraint

param: optional parameter (default to nothing)

transformation: a transformation layer (optional)

arithmetic: a arithmetic layer (optional)

aggregation: a aggregation layer (optional)

comparison: a comparison layer (optional)

CompositionalNetworks._abs_diff_val_param — Method

_abs_diff_val_param(x::T, param::T)

Return the absolute difference between x and param.

CompositionalNetworks._abs_diff_val_vars — Method

_abs_diff_val_vars(x::T, nvars::Int)

Return the absolute difference between x and the number of variables.

CompositionalNetworks._as_bitvector — Function

_as_bitvector(n::Int, max_n::Int = n)

Convert an Int to a BitVector of minimal size (relatively to max_n).

CompositionalNetworks._as_int — Method

_as_bitvector(v::AbstractVector)

Convert a BitVector into an Int.

CompositionalNetworks._contiguous_vals_minus — Method

_contiguous_vals_minus(i::Int, x::V)
_contiguous_vals_minus_rev(i::Int, x::V)

Return the difference x[i] - x[i + 1] (resp. x[i + 1] - x[i]) if positive, 0.0 otherwise. Extended method to vector with sig (x::V) are generated.

CompositionalNetworks._count_bounding_param — Method

_count_bounding_param(i::Int, x::V, param::T)

Count the number of elements bounded (not strictly) by x[i] and x[i] + param. An extended method to vector with sig (x::V, param::T) is generated.

CompositionalNetworks._count_eq — Method

_count_eq(i::Int, x::V)
_count_eq_right(i::Int, x::V)
_count_eq_left(i::Int, x::V)

Count the number of elements equal to x[i] (optionally to the right/left of x[i]). Extended method to vector x::V are generated.

CompositionalNetworks._count_eq_param — Method

_count_eq_param(i::Int, x::V, param::T)
_count_l_param(i::Int, x::V, param::T)
_count_g_param(i::Int, x::V, param::T)

Count the number of elements equal to (resp. lesser/greater than) x[i] + param. Extended method to vector with sig (x::V, param::T) are generated.

CompositionalNetworks._count_greater — Method

_count_greater(i::Int, x::V)
_count_lesser(i::Int, x::V)
_count_g_left(i::Int, x::V)
_count_l_left(i::Int, x::V)
_count_g_right(i::Int, x::V)
_count_l_right(i::Int, x::V)

Count the number of elements greater/lesser than x[i] (optionally to the left/right of x[i]). Extended method to vector with sig (x::V) are generated.

CompositionalNetworks._count_positive — Method

_count_positive(x::V)

Count the number of strictly positive elements of x.

CompositionalNetworks._euclidian_param — Method

_euclidian_param(x::T, param::T, dom_size::T2)
_euclidian(x::T, dom_size::T2)

Compute an euclidian norm , possibly weigthed by param, on a scalar.

CompositionalNetworks._exclu — Method

_exclu(layer)

Return true if the layer has mutually exclusive operations.

CompositionalNetworks._functions — Method

_functions(layer)

Access the operations of a layer. The container is ordered.

CompositionalNetworks._identity — Method

_identity(x::V) where {T <: Number,V <: AbstractVector{T}}
_identity(x::T) where T <: Number = identity(x)

Identity function. Already defined in Julia as identity, specialized for vectors and scalars.

CompositionalNetworks._layers — Method

_layers(icn)

Return the ordered layers of an ICN.

CompositionalNetworks._length — Method

_length(layer)

Return the number of operations in a layer.

CompositionalNetworks._length — Method

_length(icn)

Return the total number of operations of an ICN.

CompositionalNetworks._map_tr — Method

_map_tr(f, x)
_map_tr(f, x, param)

Return an anonymous function that applies f to all elements of x, with an optional parameter param.

CompositionalNetworks._nbits_exclu — Method

_nbits_exclu(layer)

Convert the length of an exclusive layer into a number of bits.

CompositionalNetworks._sum — Method

_sum(x::V)

Aggregate through + a vector into a single scalar.

CompositionalNetworks._sum — Method

_sum(x::W) where {T <: Number, V <: AbstractVector{T}, W <: AbstractVector{V}}
_prod(x::W) where {T <: Number, V <: AbstractVector{T}, W <: AbstractVector{V}}

Reduce k = length(x) vectors through sum/product to a single vector.

CompositionalNetworks._symbol — Method

_symbol(layer, i)

Return the i-th symbols of the operations in a given layer.

CompositionalNetworks._val_minus_param — Method

_val_minus_param(x::T, param::T)
_param_minus_val(x::T, param::T)

Return the difference x - param (resp. param - x) if positive, 0.0 otherwise.

CompositionalNetworks._val_minus_param — Method

_val_minus_param(i::Int, x::V, param::T)
_param_minus_val(i::Int, x::V, param::T)

Return the difference x[i] - param (resp. param - x[i]) if positive, 0.0 otherwise. Extended method to vector with sig (x::V, param::T) are generated.

CompositionalNetworks._val_minus_vars — Method

_val_minus_vars(x::T, nvars::Int)
_vars_minus_val(x::T, nvars::Int)

Return the difference x - nvars (resp. nvars - x) if positive, 0.0 otherwise.

CompositionalNetworks._weigths — Method

_weigths(icn)

Access the current set of weigths of an ICN.

CompositionalNetworks.aggregation_layer — Method

aggregation_layer()

Generate the layer of aggregation functions of the ICN.

CompositionalNetworks.arithmetic_layer — Method

arithmetic_layer()

Generate the layer of arithmetic functions of the ICN.

CompositionalNetworks.comparison_layer — Function

comparison_layer(nvars, dom_size, param = nothing)

Generate the layer of transformations functions of the ICN. Iff param value is set, also includes all the parametric transformation with that value.

CompositionalNetworks.compose — Method

compose(icn)

Return a function composed by some of the operations of a given ICN. Can be applied to any vector of variables.

CompositionalNetworks.lazy — Method

lazy(funcs::Function...)
lazy_param(funcs::Function...)

Generate methods extended to a vector instead of one of its components. For lazy (resp. lazy_param) a function f should have the following signature: f(i::Int, x::V) (resp. f(i::Int, x::V, param::T)).

CompositionalNetworks.transformation_layer — Function

transformation_layer(param = nothing)

Generate the layer of transformations functions of the ICN. Iff param value is set, also includes all the parametric transformation with that value.