Transformation Layer

The transformation layer of our basic ICN can be constructed using transformation_layer(param=false).

CompositionalNetworks.transformation_layer — Function

transformation_layer(param = false)

Generate the layer of transformations functions of the ICN. Iff param value is true, also includes all the parametric transformations.

Non-parametric transformations

Follows a list of the current non-parametric operations available in any transformation layer.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_identity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_eq. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_eq_left. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_eq_right. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_greater. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_lesser. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_g_left. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_l_left. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_g_right. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_l_right. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_contiguous_vals_minus. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_contiguous_vals_minus_rev. Check Documenter's build log for details.

Note that all functions are extended to a vectorized version with the lazy function.

CompositionalNetworks.lazy — Function

lazy(funcs::Function...)

Generate methods extended to a vector instead of one of its components. A function f should have the following signature: f(i::Int, x::V).

Parametric transformations

And finally a list of the parametric ones.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_eq_param. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_l_param. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_g_param. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_count_bounding_param. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_val_minus_param. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CompositionalNetworks._tr_param_minus_val. Check Documenter's build log for details.

Note that all functions are extended to a vectorized version with the lazy_param function.

CompositionalNetworks.lazy_param — Function

lazy_param(funcs::Function...)

Generate methods extended to a vector instead of one of its components. A function f should have the following signature: f(i::Int, x::V; param).